Utils parsing fix

+++ src/scaffold-theme/generate-liquid-files.ts
@@ -792,7 +792,9 @@ declare global {
       );
       if (
         fileName.includes("_layout.translations.liquid") ||
-        fileName.includes("_layout.dynamic-imports.liquid")
+        fileName.includes("_layout.dynamic-imports.liquid") ||
+        /^replo/gi.test(fileName) ||
+        /^_preset/gi.test(fileName)
       ) {
         return;
       }
@@ -812,6 +814,9 @@ declare global {
         Object.entries(sources.sectionPresetSchemas).find(
           ([key, val]) => !val.disabled && fileName === `preset__${toSnakeCase(key)}.liquid`
         );
+      if (/^replo/gi.test(fileName) || /^_preset/gi.test(fileName)) {
+        return;
+      }
       if (!targetFile) {
         deleteFile(path.join(process.cwd(), file));
       }
